Move args test to new class
